Moved checkDownloadedFileEquals out into a shared helper file.
